requirements.txt
anyio==3.2.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
brotlicffi==1.0.9.2
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.5
django-filter==2.1.0
django-guardian==2.0.0
django-heroku==0.3.1
djangorestframework==3.10.1
entrypoints==0.3
evdev==1.3.0
filelock==3.0.12
gpg===1.13.1-unknown
h11==0.12.0
h2==3.2.0
hpack==3.0.0
html5lib==1.0.1
httpcore==0.13.6
httplib2==0.14.0
httpx==0.18.2
hyperframe==5.2.0
idna==2.10
itypes==1.1.0
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Markdown==3.1.1
MarkupSafe==2.0.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
peewee==3.14.4
pexpect==4.6.0
Pillow==8.3.0
protobuf==3.6.1
psycopg2==2.8.4
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
PySocks==1.6.8
pyTelegramBotAPI==3.7.9
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-telegram-bot-calendar==1.0.5
python-xlib==0.23
pytz==2021.1
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.2.0
soupsieve==1.9.5
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
tbomb==2.1.2
TelegramBotAPI==0.3.2
termcolor==1.1.0
torbrowser-launcher==0.3.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
virtualenv==20.4.7
wadllib==1.3.3
webencodings==0.5.1
whitenoise==5.3.0
xkit==0.0.0
